Adds a host tag to tag metrics with `"production"` or `"dev"`.
When practicing the presentation, the metrics looked crazy when an instance of the appliation was running on a local dev machine.

To avoid the chance of this happening during the presentation, this commit adds a `host` tag and uses the environment to set it.

We will want to update our DataDog dashboard to filter for `host` to be `"production"` as well.

This may make local dev a little tickier when figuring out metrics, but we should be able to use the Metrics Explorer or temporairily change the charts in the dashboard.