Refactors how we work with metrics.
Currently, we are initializing a new "instance" of our metrics connection to DataDog in each file that we are using metrics. The issue with this is that if we want to change how we are sending the metrics (e.g., new prefix or adding host), we have to change it in all of the spots.

This commit extract the initaliziation logic into its own module and uses that new module whenever we need metrics.