Reduce topping-list max-height.
A height of `100vh` didn't account for the fact that the `navbar` and
additional `margin`/`padding` would be taking up space.

When full-screen this shouldn't change anything, but on smaller screens
it _should_ prevent the need to scroll beyond the right side menu bar.
If this were a production thing we'd would want to more throughly test
this.